Improve performance for UB check with identical src/tgt
We often have identical IR for src and tgt. Z3 is reasonably
good at seeing this and quickly return UNSAT. But it fails doing
this for some functions, and we in worst case end up with a
timeout for each pass (i.e., taking several hours to compile).
This issue goes away when we add a redundant src.ub != tgt.ub
check.