Improve pointer validity check
The pointer validity checks added in the last commits are not
correct. There are several issues (I will discuss this in a blog
post soon), but the short version is that it is not wrong to have
invalid pointers in GIMPLE. One way we can get such pointers is

  struct S {int *p1; int *p2;};
  int *foo(struct S *p, int c) {
    if (c)
      return p->p1;
    else
      return p->p2;
  }

Assume this is called as foo(&s, 1) where s.p1 is valid and s.p2
is invalid (e.g. a freed pointer, or uninitialized field). This
is OK as s.p2 is never read, but phiopt may transform this to

  int *foo(struct S *p, int c) {
    int *t1 = p->p1;
    int *t2 = p->p2;
    if (c)
      return t1;
    else
      return t2;
  }

to make it possible to vectorize the load instructions.

So we must allow invalid pointers. But using such pointers is
not allowed. We therefore handle this in essentially the same
way as uninitialized data and check it in get_tree_as_smt.

We must also do more detailed checks for some of the operations
(for example, check that the whole structure fits in the memory
block when accessing one element). This will be done in a later
commit.