Do not check return that does not return a value
C allows returning without a value, even in functions that are
declared to return a value, such as

  int foo(int a) {
    if (a)
      return;
    return 42;
  }

The returned value is undefined, but that is OK as long as the
caller is not using the value. This means that we should not
compare the returned value for the returns without value, and
we should verify that the tgt function return a value for all
cases where src returns a value.